modetest/requirements.txt
ymodem==1.5.1
pyserial==3.5 